Use `NODE_BINARY` from `.xcode.env` when running packager from Xcode (#34121)
Summary:
Before this change, during the "Start Packager" Xcode build step that runs the packager, `packager.sh` was using my system node version and not the one from `nvm` that is meant to be used with the project.

## Changelog

[iOS] [Fixed] - Use `NODE_BINARY` from `.xcode.env` when running packager from Xcode

Pull Request resolved: https://github.com/facebook/react-native/pull/34121

Test Plan: Perform a build using Xcode and confirm that the packager is using the correct version of node.

Reviewed By: cortinico

Differential Revision: D37746951

Pulled By: cipolleschi

fbshipit-source-id: de697c27fe86ce65e8e3646cb30309ecc7f6c247